requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
dnspython==2.4.2
pydantic==2.5.0
python-multipart==0.0.6
